добавил взаимодействие с БД item и user

+++ src/main/java/ru/practicum/shareit/item/ItemController.java
@@ -14,28 +14,28 @@ public class ItemController {
     private final ItemService itemService;
 
     @GetMapping("/{itemId}")
-    public ItemDto get(@RequestHeader("X-Sharer-User-Id") Integer userId, @PathVariable Integer itemId) {
+    public ItemDto get(@RequestHeader("X-Sharer-User-Id") Long userId, @PathVariable Long itemId) {
         return itemService.getItem(userId, itemId);
     }
 
     @GetMapping("")
-    public List<ItemDto> getItems(@RequestHeader("X-Sharer-User-Id") Integer userId) {
+    public List<ItemDto> getItems(@RequestHeader("X-Sharer-User-Id") Long userId) {
         return itemService.getItems(userId);
     }
 
     @PostMapping
-    public ItemDto add(@RequestHeader("X-Sharer-User-Id") Integer userId,
+    public ItemDto add(@RequestHeader("X-Sharer-User-Id") Long userId,
                        @RequestBody ItemDto item) {
         return itemService.addItem(userId, item);
     }
 
     @PatchMapping("/{itemId}")
-    public ItemDto update(@Valid  @PathVariable Integer itemId, @RequestHeader("X-Sharer-User-Id") Integer userId, @RequestBody ItemDto item) {
+    public ItemDto update(@Valid  @PathVariable Long itemId, @RequestHeader("X-Sharer-User-Id") Long userId, @RequestBody ItemDto item) {
         return itemService.updateItem(userId, itemId, item);
     }
 
     @GetMapping("/search")
-    public List<ItemDto> searchItems(@RequestParam("text") String searchText, @RequestHeader("X-Sharer-User-Id") Integer userId) {
+    public List<ItemDto> searchItems(@RequestParam("text") String searchText, @RequestHeader("X-Sharer-User-Id") Long userId) {
         return itemService.getAvailableItems(searchText, userId);
     }
 }
