+++ src/main/java/ru/practicum/shareit/user/UserService.java
@@ -3,11 +3,11 @@ package ru.practicum.shareit.user;
 import ru.practicum.shareit.user.dto.UserDto;
 
 public interface UserService {
-    UserDto getUser(Integer userId);
+    UserDto getUser(Long userId);
 
     UserDto addUser(UserDto userDto);
 
-    UserDto updateUser(Integer userId, UserDto userDto);
+    UserDto updateUser(Long userId, UserDto userDto);
 
-    UserDto deleteUser(Integer userId);
+    UserDto deleteUser(Long userId);
 }
